Move FakeRequestBuilder from the auth package to support
